refactor: split function file
