+++ functions/src/admin.ts
@@ -1,0 +1,4 @@
+import * as admin from "firebase-admin";
+
+export const db = admin.database();
+export const firestore = admin.firestore();
